Partially Completed Animation Controller
The animation controller uses a 2D Blend Tree. Mostly done, just have to tie strafing to whether or not the player is using the right stick. Forgot my charger so it is hard to test with all the lag. Noticed a couple strange happenings with the movement, but I hope it's just the lag.